Added some more tests for polyphasegenetic and the missing files for the previous tests.
