+++ tests/data/polyphasegenetic.ped1.txt
@@ -1,0 +1,5 @@
+Parent_A Parent_B p1
+Parent_A Parent_B p2
+Parent_A Parent_B p3
+Parent_A Parent_B p4
+Parent_A Parent_B p4
