+++ tests/test_variantselection.py
@@ -3,70 +3,6 @@ from whatshap.variantselection import VariantInfo, compute_phasable_variants
 from whatshap.vcf import VcfReader
 
 
-def test_read_parent_vcf():
-    tables = list(
-        VcfReader(
-            "tests/data/polyphasegenetic.test.parents.vcf",
-            indels=True,
-            genotype_likelihoods=False,
-            ploidy=4,
-            mav=True,
-        )
-    )
-    assert len(tables) == 1
-    table = tables[0]
-
-    assert table.chromosome == "StSOLv1.1ch03"
-    assert len(table.variants) == 135
-    assert table.samples == ["Parent_A", "Parent_B"]
-
-    assert table.variants[0].reference_allele == "C"
-    assert table.variants[0].alternative_allele == "T"
-    assert table.variants[33].reference_allele == "ACC"
-    assert table.variants[33].alternative_alleles == ["A", "AC"]
-    assert table.variants[34].reference_allele == "C"
-    assert table.variants[34].alternative_alleles == ["*", "T"]
-
-
-def test_read_progeny_vcf():
-    tables = list(
-        VcfReader(
-            "tests/data/polyphasegenetic.test.progeny.vcf.gz",
-            indels=True,
-            genotype_likelihoods=False,
-            ploidy=4,
-            mav=True,
-            allele_depth=True,
-        )
-    )
-    assert len(tables) == 1
-    table = tables[0]
-
-    assert table.chromosome == "StSOLv1.1ch03"
-    assert len(table.variants) == 198
-    assert len(table.samples) == 64
-    assert table.samples == ["Progeny_{}".format(i) for i in range(1, 65)]
-
-    assert table.variants[0].reference_allele == "C"
-    assert table.variants[0].alternative_allele == "T"
-    assert table.variants[19].reference_allele == "T"
-    assert table.variants[19].alternative_alleles == ["C", "A"]
-    assert table.variants[71].reference_allele == "AGT"
-    assert table.variants[71].alternative_alleles == ["AGGT", "*"]
-
-    ad1 = table.allele_depths_of("Progeny_1")
-    ad2 = table.allele_depths_of("Progeny_2")
-    ad3 = table.allele_depths_of("Progeny_3")
-    ad4 = table.allele_depths_of("Progeny_4")
-
-    assert ad1[0] == (1, 4)
-    assert ad2[0] == (9,)
-    assert ad3[0] == (7,)
-    assert ad4[0] == (4, 3)
-    assert ad1[19] == (7,)
-    assert table.allele_depths_of("Progeny_62")[44] == (3, 1, 1)
-
-
 def test_variantinfo_1():
     vi = VariantInfo([(1, 0)])
     assert vi.get_phasable() == []
